NEXUS-481 Started Port to TestNG,
Problems:
I only converted the non-secure suite,  It works mostly 10 errors.  The suite starts, and if you run 1 or 10 classes you get the same flow, (which is better then JUnit)
be the order is undetermined, unlike the order of the tests in JUnit, TestNG supports ordering the Test via Annotations, but that is not the real problem, a few of the tests seem to depend on the others in some way.

Also, when the Nexus636EvictUnusedProxiedTaskTest test is included in the suite the whole thing fails, I don't know what that is. But this is not a critical issue right now, and i need to move on... So, i am putting this on a branch.

TODO:
Convert Secuirty Test Suite.
    (Something needs to enable the security flag)
Make tests not depend on each other



git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/branches@1864 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab